docs: App.jsx의 내용 InitScreen.jsx로 옮기고 그에 맞게 코드 수정
